requirements.txt
pytesseract
pyautogui
Pillow
pyinstaller
customtkinter
pygments
pywin32
psutil
keyboard
pytesseract